docs + renaming
