+++ DFT_1d/ks_inversion.py
@@ -15,7 +15,7 @@ External Potentials
 
 import numpy as np
 from single_electron import SparseEigenSolver
-from formula import IntegralTool, DerivativeTool
+from utils import IntegralTool, DerivativeTool
 from ext_potentials import exp_hydrogenic
 import functools
 import scipy
