+++ DFT_1d/utils.py
